Code review #51
Moved all podman structs to new file structs.go
